Add files via upload
There's an obvious problem with our way of computing loss. As it stands, we are computing the jacobian of our function at each step and each time we do so, @tf.function has to systematically re-compute the computational graph for the operations our MLP does. I did not find any solutions to this problem unfortunately.